web-history-extension/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
pydantic==2.7.2
python-dotenv==1.0.1
requests==2.31.0
numpy==1.24.3
faiss-cpu==1.7.4
markitdown==0.1.0
google-generativeai==0.3.1
python-multipart==0.0.6
fastapi_mcp==0.1.0
